Refactor the tests by creating different folders
